Add files via upload
Note the 3 python files are required in this project. The first file, AddDataToDatabase.py is used to store the profiles of stored faces in firebase storage.
The second file, Encoder.py is used to encode all the characteristics and functionalities used for face recognition and Attendance Monitoring.
The Third file, main.py is the primary python file which executes the program.